Refactor scheduling and solenoid features by removing obsolete widgets and implementing new structure
- Deleted ScheduleItem and ScheduleList widgets to streamline scheduling UI.
- Removed SolenoidScreen and related widgets to enhance solenoid management.
- Introduced new entities for Schedule and Solenoid to improve data handling.
- Added AddScheduleSheet and EditScheduleSheet for better scheduling management.
- Implemented BuildDayChip for day selection in scheduling.
- Updated SolenoidItem and SolenoidList for improved solenoid display.
- Created SolenoidSettingSheet and SolenoidSettingSheetChose for solenoid configuration.
