+++ lib/src/features/scheduling/domain/entities/schedule.dart
@@ -1,0 +1,10 @@
+// class Schedule {
+//   final int id;
+//   final int modulId;
+//   final String label;
+//   final int duration;
+//   final String time;
+//   final bool isActive;
+//   final DateTime createdAt;
+//   final DateTime updatedAt;
+// }
