+++ lib/src/features/scheduling/presentation/widgets/schedule_widgets/schedule_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pak_tani/src/core/theme/app_theme.dart';
 import 'package:pak_tani/src/core/widgets/my_switch.dart';
-import 'package:pak_tani/src/features/scheduling/presentation/widgets/edit_schedule_sheet.dart';
+import 'package:pak_tani/src/features/scheduling/presentation/widgets/schedule_widgets/edit_schedule_sheet.dart';
 
 class ScheduleItem extends StatelessWidget {
   final int duration;
