+++ lib/src/features/scheduling/presentation/widgets/solenoid_widgets/selenoid_item.dart
